+++ scripts/macros/spells/bless.js
@@ -11,19 +11,21 @@ export async function bless(args) {
 	nearbyTargets.push(token);
 	let buttons = [{label: 'Apply to Selected', value: true},{label: 'Cancel', value: false}];
 	const maxTargets = lastArg.castData.castLevel+2;
-	let chosenTargets = await ggHelpers.selectTarget(`Bless Targets - (Max: ${maxTargets})`, buttons, nearbyTargets, false, true);
+	let chosenTargets = await ggHelpers.selectTarget(`Bless Targets - (Max: ${maxTargets})`, buttons, nearbyTargets, true);
 	if (chosenTargets) {
 		if (!chosenTargets.buttons) return;
 		let numSelectedTargets = chosenTargets.inputs.filter(val => val !== false).length;
 		while(numSelectedTargets > maxTargets) {
-			chosenTargets = await ggHelpers.selectMultipleTargets(`Chose Too Many Targets! (Max: ${maxTargets})`, buttons, nearbyTargets, false, true);
+			chosenTargets = await ggHelpers.selectTarget(`Chose Too Many Targets! (Max: ${maxTargets})`, buttons, nearbyTargets, true);
 			if (!chosenTargets) return;
 			if (!chosenTargets.buttons) return;
 			numSelectedTargets = chosenTargets.inputs.filter(val => val !== false).length;
 		}
-		for (let target of chosenTargets.inputs) {
-			if (target) {
-				let targetToken = canvas.tokens.get(target);
+		console.log(chosenTargets);
+		for (let i = 0; i < chosenTargets.inputs.length; i++) {
+			if (chosenTargets.inputs[i]) {
+				let targetToken = nearbyTargets[i];
+				console.log(targetToken);
 				targetToken.setTarget(true, { user: game.user, releaseOthers: false });
 			}
 		}